+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 ## [1.1.0-14] - 16/8/21
 ### Changes
-- intro section using css background gradient 
-- 
+- intro section using css background gradient
+- shadow to sidenav
+- scrollbar color
+- video rows spacing
+  
+### Fixes
+- fix embed for hero video
+
+
 ## [1.1.0-13] - 16/8/21
 ### Fixes
 - made slant svgs responsive